[MSHARED-860] Move createSymbolicLink() utility to FileUtils
