+++ src/test/java/org/apache/maven/shared/utils/io/Java7SupportTest.java
@@ -43,20 +43,4 @@ public class Java7SupportTest
         }
     }
 
-    @Test
-    public void createAndReadSymlink()
-        throws Exception
-    {
-        assumeThat( System.getProperty( "os.name" ), not( startsWith( "Windows" ) ) );
-        File file = new File( "target/fzz" );
-        if ( Java7Support.isAtLeastJava7() )
-        {
-            Java7Support.createSymbolicLink(  file, new File("../target") );
-
-            final File file1 = Java7Support.readSymbolicLink( file );
-            assertEquals( "target", file1.getName());
-            Java7Support.delete(  file );
-        }
-    }
-
 }
